+++ packages/playground/src/Bootstrap/Popover.js
@@ -1,4 +1,4 @@
-import { css } from 'glamor';
+import { css } from '@emotion/css';
 import React, { useCallback, useEffect, useRef } from 'react';
 
 const ROOT_CSS = css({